submodule: gracefully handle links in module paths
Due to security concerns the commit e8d060894448 (submodule: require the
submodule path to contain directories only, 2024-03-26) introduced some
checks and hard `exit(128)` calls for the case that a submodule path
contains symbolic links. This change was motivated by CVE-2024-32002,
which was discovered shortly before. Consequently, git currently aborts
actions of which some could have been continued by not processing the
faulty submodule.

As an example, one of these actions that may be aborted is showing a
diff with diff.submodule set to `log` or `diff`. In a repository where a
submodule was moved and its old path replaced with a symlink pointing to
the new location, a git-show on an older commit that touches the
submodule before the move will have only partial output. Other examples
for actions that are now continued are fetching and pushing with their
respective recurseSubmodules turned on.

Handle these cases more gracefully by returning an error value instead
of dying, if possible.

CC: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Julian Prein <julian@druck.dev>

Following code was touched:

TODO: give explanation for each one

  - `push_submodule()`:  can return instead of exiting so that
    `push_unpushed_submodules()` simply skips the problematic module and
    continues pushing the remaining modules.
  - `fetch_task_create()`
  - `bad_to_remove_submodule()`
  - `submodule_to_gitdir()` has multiple uses (TODO). One of them is in
    `open_submodule()` which is used when diff.submodule is set to log or diff.

old:

WIP: Gracefully exit when a submodule path contains a link

Most important one is `submodule_to_gitdir` as this is used in git-diff
when diff.submodule is set to log. When a submodule was moved and
replaced by a symlink, all old commits can't be shown currently. With
this commit, git-show works by practically falling back to
`diff.submodule=short`, i.e. it still prints the error message but
displays the diff and for each submodule the revision range with
`(commits not present)` or `(new submodule)` without the log. The same
behaviour is emitted when the submodule was simply moved to a new path
(TODO: which could maybe also be done better).